MenuProduct에서 Product 직접 의존하고 Validator 제거
