+++ src/main/java/kitchenpos/menu/application/MenuService.java
@@ -2,50 +2,45 @@ package kitchenpos.menu.application;
 
 import java.util.List;
 import kitchenpos.menu.application.dto.MenuCreateRequest;
-import kitchenpos.menu.application.dto.MenuCreateRequest.MenuProductInfo;
 import kitchenpos.menu.application.dto.MenuResponse;
 import kitchenpos.menu.domain.Menu;
 import kitchenpos.menu.domain.MenuGroupRepository;
 import kitchenpos.menu.domain.MenuProduct;
 import kitchenpos.menu.domain.MenuRepository;
-import kitchenpos.menu.domain.MenuValidator;
+import kitchenpos.product.domain.ProductRepository;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 @Service
 public class MenuService {
 
+    private final ProductRepository productRepository;
     private final MenuGroupRepository menuGroupRepository;
     private final MenuRepository menuRepository;
-    private final MenuValidator menuValidator;
 
     public MenuService(
+            ProductRepository productRepository,
             MenuGroupRepository menuGroupRepository,
-            MenuRepository menuRepository,
-            MenuValidator menuValidator
+            MenuRepository menuRepository
     ) {
+        this.productRepository = productRepository;
         this.menuGroupRepository = menuGroupRepository;
         this.menuRepository = menuRepository;
-        this.menuValidator = menuValidator;
     }
 
     @Transactional
-    public MenuResponse create(final MenuCreateRequest request) {
-        List<MenuProduct> menuProducts = menuProducts(request.getMenuProducts());
-        Menu menu = Menu.create(
+    public MenuResponse create(MenuCreateRequest request) {
+        List<MenuProduct> menuProducts = request.getMenuProducts()
+                .stream()
+                .map(it -> new MenuProduct(productRepository.getById(it.getProductId()), it.getQuantity()))
+                .toList();
+        Menu menu = new Menu(
                 request.getName(),
                 request.getPrice(),
                 menuGroupRepository.getById(request.getMenuGroupId()),
-                menuProducts,
-                menuValidator
+                menuProducts
         );
         return MenuResponse.from(menuRepository.save(menu));
-    }
-
-    private List<MenuProduct> menuProducts(List<MenuProductInfo> menuProducts) {
-        return menuProducts.stream()
-                .map(it -> new MenuProduct(it.getProductId(), it.getQuantity()))
-                .toList();
     }
 
     public List<MenuResponse> list() {
