+++ src/main/java/kitchenpos/menu/domain/Menu.java
@@ -5,6 +5,7 @@ import static javax.persistence.GenerationType.IDENTITY;
 
 import java.math.BigDecimal;
 import java.util.List;
+import java.util.Objects;
 import javax.persistence.Entity;
 import javax.persistence.FetchType;
 import javax.persistence.GeneratedValue;
@@ -12,6 +13,7 @@ import javax.persistence.Id;
 import javax.persistence.JoinColumn;
 import javax.persistence.ManyToOne;
 import javax.persistence.OneToMany;
+import kitchenpos.product.domain.Product;
 
 @Entity
 public class Menu {
@@ -33,22 +35,32 @@ public class Menu {
     protected Menu() {
     }
 
-    private Menu(String name, BigDecimal price, MenuGroup menuGroup, List<MenuProduct> menuProducts) {
+    public Menu(String name, BigDecimal price, MenuGroup menuGroup, List<MenuProduct> menuProducts) {
+        validateCreate(price, menuGroup, menuProducts);
         this.name = name;
         this.price = price;
         this.menuGroup = menuGroup;
         this.menuProducts = menuProducts;
     }
 
-    public static Menu create(
-            String name,
-            BigDecimal price,
-            MenuGroup menuGroup,
-            List<MenuProduct> menuProducts,
-            MenuValidator menuValidator
-    ) {
-        menuValidator.validateCreate(price, menuGroup, menuProducts);
-        return new Menu(name, price, menuGroup, menuProducts);
+    private void validateCreate(BigDecimal price, MenuGroup menuGroup, List<MenuProduct> menuProducts) {
+        if (menuGroup == null) {
+            throw new InvalidMenuException("메뉴는 메뉴 그룹에 속해야 합니다.");
+        }
+        if (menuProducts.isEmpty()) {
+            throw new InvalidMenuException("메뉴에는 최소 1개의 상품이 속해야 합니다.");
+        }
+        if (Objects.isNull(price) || price.compareTo(BigDecimal.ZERO) < 0) {
+            throw new InvalidMenuException("메뉴의 가격은 0원 이상이어야 합니다.");
+        }
+        BigDecimal sum = BigDecimal.ZERO;
+        for (MenuProduct menuProduct : menuProducts) {
+            Product product = menuProduct.getProduct();
+            sum = sum.add(product.getPrice().multiply(BigDecimal.valueOf(menuProduct.getQuantity())));
+        }
+        if (price.compareTo(sum) > 0) {
+            throw new InvalidMenuException("메뉴의 가격은 메뉴에 포함된 상품들의 합 이하여야 합니다.");
+        }
     }
 
     public Long getId() {
