+++ src/main/java/kitchenpos/menu/domain/MenuProduct.java
@@ -3,8 +3,12 @@ package kitchenpos.menu.domain;
 import static javax.persistence.GenerationType.IDENTITY;
 
 import javax.persistence.Entity;
+import javax.persistence.FetchType;
 import javax.persistence.GeneratedValue;
 import javax.persistence.Id;
+import javax.persistence.JoinColumn;
+import javax.persistence.ManyToOne;
+import kitchenpos.product.domain.Product;
 
 @Entity
 public class MenuProduct {
@@ -12,14 +16,18 @@ public class MenuProduct {
     @Id
     @GeneratedValue(strategy = IDENTITY)
     private Long seq;
-    private Long productId;
+
+    @ManyToOne(fetch = FetchType.LAZY)
+    @JoinColumn(name = "product_id")
+    private Product product;
+
     private long quantity;
 
     protected MenuProduct() {
     }
 
-    public MenuProduct(Long productId, long quantity) {
-        this.productId = productId;
+    public MenuProduct(Product product, long quantity) {
+        this.product = product;
         this.quantity = quantity;
     }
 
@@ -27,8 +35,8 @@ public class MenuProduct {
         return seq;
     }
 
-    public Long getProductId() {
-        return productId;
+    public Product getProduct() {
+        return product;
     }
 
     public long getQuantity() {
