+++ src/test/java/kitchenpos/menu/MenuTest.java
@@ -2,17 +2,14 @@ package kitchenpos.menu;
 
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 import static org.junit.jupiter.api.Assertions.assertDoesNotThrow;
-import static org.mockito.ArgumentMatchers.any;
-import static org.mockito.BDDMockito.willDoNothing;
-import static org.mockito.BDDMockito.willThrow;
-import static org.mockito.Mockito.mock;
 
 import java.math.BigDecimal;
 import java.util.List;
 import kitchenpos.menu.domain.InvalidMenuException;
+import kitchenpos.menu.domain.Menu;
 import kitchenpos.menu.domain.MenuGroup;
 import kitchenpos.menu.domain.MenuProduct;
-import kitchenpos.menu.domain.MenuValidator;
+import kitchenpos.product.domain.Product;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.DisplayNameGeneration;
 import org.junit.jupiter.api.DisplayNameGenerator.ReplaceUnderscores;
@@ -24,38 +21,65 @@ import org.junit.jupiter.api.Test;
 @DisplayNameGeneration(ReplaceUnderscores.class)
 class MenuTest {
 
-    private final MenuValidator menuValidator = mock(MenuValidator.class);
-
     @Nested
     class 생성_시 {
 
         @Test
-        void 검증시_오류가_존재하면_생성되지_않는다() {
-            // given
-            willThrow(new InvalidMenuException())
-                    .given(menuValidator)
-                    .validateCreate(any(), any(), any());
-
+        void 메뉴가_메뉴_그룹에_속하지_않는다면_예외() {
             // when & then
-            assertThatThrownBy(() -> menuValidator.validateCreate(
+            assertThatThrownBy(() -> new Menu(
+                    "말랑메뉴",
                     BigDecimal.valueOf(1000),
                     null,
-                    List.of(new MenuProduct(1L, 2L))
-            )).isInstanceOf(InvalidMenuException.class);
+                    List.of(new MenuProduct(new Product("말랑", BigDecimal.valueOf(1000)), 2L))
+            )).isInstanceOf(InvalidMenuException.class)
+                    .hasMessage("메뉴는 메뉴 그룹에 속해야 합니다.");
         }
 
         @Test
-        void 검증시_오류가_없으면_생성된다() {
-            // given
-            willDoNothing()
-                    .given(menuValidator)
-                    .validateCreate(any(), any(), any());
-
+        void 메뉴에_속한_상품이_없으면_예외() {
             // when & then
-            assertDoesNotThrow(() -> menuValidator.validateCreate(
+            assertThatThrownBy(() -> new Menu(
+                    "말랑메뉴",
                     BigDecimal.valueOf(1000),
-                    new MenuGroup("말랑"),
-                    List.of(new MenuProduct(1L, 2L))
+                    new MenuGroup("메뉴그룹 1"),
+                    List.of()
+            )).isInstanceOf(InvalidMenuException.class)
+                    .hasMessage("메뉴에는 최소 1개의 상품이 속해야 합니다.");
+        }
+
+        @Test
+        void 메뉴의_가격이_0원_미만이라면_예외() {
+            // when & then
+            assertThatThrownBy(() -> new Menu(
+                    "말랑메뉴",
+                    BigDecimal.valueOf(-1),
+                    new MenuGroup("메뉴그룹 1"),
+                    List.of(new MenuProduct(new Product("말랑", BigDecimal.valueOf(1000)), 2L))
+            )).isInstanceOf(InvalidMenuException.class)
+                    .hasMessage("메뉴의 가격은 0원 이상이어야 합니다.");
+        }
+
+        @Test
+        void 메뉴의_가격이_메뉴에_속한_상품들의_합보다_크면_예외() {
+            // when & then
+            assertThatThrownBy(() -> new Menu(
+                    "말랑메뉴",
+                    BigDecimal.valueOf(2001),
+                    new MenuGroup("메뉴그룹 1"),
+                    List.of(new MenuProduct(new Product("말랑", BigDecimal.valueOf(1000)), 2L))
+            )).isInstanceOf(InvalidMenuException.class)
+                    .hasMessage("메뉴의 가격은 메뉴에 포함된 상품들의 합 이하여야 합니다.");
+        }
+
+        @Test
+        void 예외_상황이_없는_경우() {
+            // when & then
+            assertDoesNotThrow(() -> new Menu(
+                    "말랑메뉴",
+                    BigDecimal.valueOf(200),
+                    new MenuGroup("메뉴그룹 1"),
+                    List.of(new MenuProduct(new Product("말랑", BigDecimal.valueOf(1000)), 2L))
             ));
         }
     }
